requirements.txt
flask==2.3.3
openai==0.28.1
requests==2.31.0
pillow==10.0.1
python-dotenv==1.0.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
 youtube-transcript-api==0.6.2
gunicorn==21.2.0